fix a crash (probably)

+++ main.c
@@ -119,7 +119,7 @@ int main(int argv, char **argc) {
 	Term expr[EXPR_LEN];
 	float goal = atof(argc[1]);
 
-	int prods[EXPR_LEN] = {};
+	unsigned long long prods[EXPR_LEN] = {};
 	for (int i = 0; i < EXPR_LEN; ++i) {
 		prods[i] = 1;
 		for (int j = 0; j < i; j++) {
